move debug into a module
